requirements.txt
playwright
playwright-stealth
requests
asyncio
datetime
python-telegram-bot
